+++ CHANGELOG.md
@@ -852,4 +852,4 @@ All notable changes to this project will be documented in this file.
 ## [2.1.3] - 24.11.2023
 
 - fix prepare error on execute not showing line
-- add preview grey hack output
+- add preview grey hack output (WIP)